mavlink: move UTM_GLOBAL_POSITION to separate stream header
